+++ src/impls/is-object/__tests__/is-object.spec.ts
@@ -1,0 +1,9 @@
+import assert from 'static-type-assert';
+
+import { isObject } from 'iter-tools-es';
+
+declare const value: unknown;
+
+if (isObject(value)) {
+  assert<Record<string, any>>(value);
+}