+++ src/impls/is-object/is-object.d.ts
@@ -1,0 +1,3 @@
+declare function isObject(value: any): value is Record<string, any>;
+
+export { isObject };