+++ src/impls/not-object/not-object.d.ts
@@ -1,0 +1,3 @@
+declare function notObject<T>(value: T | Record<string, any>): value is T;
+
+export { notObject };