+++ src/impls/not-object/not-object.js
@@ -1,0 +1,3 @@
+export function notObject(value) {
+  return typeof value !== 'object' || value === null;
+}